pyproject.toml
[project]
name = "mcp-server-web-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "fastmcp>=2.13.0.2",
    "groq>=0.33.0",
    "python-dotenv>=1.2.1",
    "trafilatura>=2.0.0",
]
